Split routers into distinct files
